notifyd/requirements.txt
discord.py
python-dotenv
requests
discord.py
